+++ Sources/SuperwallKit/StoreKit/Transactions/TransactionManager.swift
@@ -61,7 +61,7 @@ final class TransactionManager {
     let product: StoreProduct
 
     switch purchaseSource {
-    case .internal(let productId, _):
+    case .internal(let productId, _, _):
       guard let storeProduct = await storeKitManager.productsById[productId] else {
         Logger.debug(
           logLevel: .error,
@@ -126,7 +126,7 @@ final class TransactionManager {
         case .observing:
           break
         case .purchasing(let purchaseSource):
-          if case let .internal(_, paywallViewController) = purchaseSource {
+          if case let .internal(_, paywallViewController, _) = purchaseSource {
             await paywallViewController.togglePaywallSpinner(isHidden: true)
           }
         }
@@ -460,7 +460,7 @@ final class TransactionManager {
     }
 
     switch source {
-    case .internal(_, let paywallViewController):
+    case .internal(_, let paywallViewController, _):
       Logger.debug(
         logLevel: .debug,
         scope: .transactions,
@@ -560,7 +560,7 @@ final class TransactionManager {
     let shouldTrackTransactionStart = !(purchaseManager.isUsingSK2 && isObserved)
 
     switch purchaseSource {
-    case .internal(_, let paywallViewController):
+    case .internal(_, let paywallViewController, _):
       Logger.debug(
         logLevel: .debug,
         scope: .transactions,
@@ -634,7 +634,7 @@ final class TransactionManager {
     }
 
     switch source {
-    case .internal(_, let paywallViewController):
+    case let .internal(_, paywallViewController, shouldDismiss):
       guard let product = await coordinator.product else {
         return
       }
@@ -659,11 +659,17 @@ final class TransactionManager {
       await trackTransactionDidSucceed(transaction)
 
       let superwallOptions = factory.makeSuperwallOptions()
-      if superwallOptions.paywalls.automaticallyDismiss {
+      let shouldDismissPaywall = superwallOptions.paywalls.automaticallyDismiss && shouldDismiss
+      if shouldDismissPaywall {
         await Superwall.shared.dismiss(
           paywallViewController,
           result: .purchased(product)
         )
+      }
+      if !shouldDismissPaywall {
+        await MainActor.run {
+          paywallViewController.togglePaywallSpinner(isHidden: true)
+        }
       }
     case .purchaseFunc,
       .observeFunc:
@@ -705,7 +711,7 @@ final class TransactionManager {
     }
 
     switch source {
-    case .internal(_, let paywallViewController):
+    case .internal(_, let paywallViewController, _):
       Logger.debug(
         logLevel: .debug,
         scope: .transactions,
@@ -765,7 +771,7 @@ final class TransactionManager {
     }
 
     switch source {
-    case .internal(_, let paywallViewController):
+    case .internal(_, let paywallViewController, _):
       Logger.debug(
         logLevel: .debug,
         scope: .transactions,
@@ -895,7 +901,7 @@ final class TransactionManager {
     let eventSource: InternalSuperwallEvent.Transaction.Source
     let trialEndDate = product.trialPeriodEndDate
     switch source {
-    case .internal(_, let paywallViewController):
+    case .internal(_, let paywallViewController, _):
       paywallInfo = await paywallViewController.info
       eventSource = .internal
       await paywallViewController.webView.messageHandler